TextGenerationPipeline.
